[FlexNG] Use block % resolution size when finding image preferred widths
Flex uses image preferred widths as flex-basis. When an image has
max-height:blah%, that can affect its preferred width. FlexNG wasn't
passing the percent resolution height to legacy for this calculation,
and the % max-height was being treated as auto.

ComputeMinAndMaxContentContribution was already doing this for replaced
elements, so piggy-back on that code.

Bug: 1107604
Change-Id: Icecb312ce8aee22cad6e4b1925aab77ecbd3c2ef